requirements.txt
amqp==5.1.1
asgiref==3.5.0
billiard==3.6.4.0
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==4.0.4
django-compat==1.0.15
django-rest-framework==0.1.0
djangorestframework==3.13.1
kombu==5.2.4
prompt-toolkit==3.0.29
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
vine==5.0.0
wcwidth==0.2.5
